refactor: move all the methods that are being used for adding validation to a field to separate util methods
Refs: #DL-5289
